# server/app/src/tsoa-build/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"AuthErrorResponseBody": {
				"$ref": "#/components/schemas/ErrorResponseBody_AuthErrorResponseContent.AuthErrorBodyMessageT_",
				"description": "The response body for an {@link AuthError}"
			},
...
			"ValidationErrorResponseBody": {
				"$ref": "#/components/schemas/ErrorResponseBody_ValidationErrorResponseContent.ValidationErrorMessageT_",
				"description": "Represents the response body for a validation error\n\nUsed by internal {@link QueryValidationError} and {@link ClassValidationError},\nas well as `sequelize`.{@link ValidationError} and `tsoa`.`ValidateError`"
			},
...
			"OrderByFieldValues": {
				"description": "Possible values for the filter query paramater on controller routes",
				"enum": [
					"createdAt",
					"updatedAt",
					"-createdAt",
					"-updatedAt"
				],
				"type": "string"
			},
...
			"NotFoundResponseBody": {
				"$ref": "#/components/schemas/ErrorResponseBody_NotFoundResponseContent.NotFoundMessageT_",
				"description": "The type of {@link ErrorResponseBody} for not found responses"
			},
...
			"AppModelIdT": {
				"type": "number",
				"format": "double",
				"description": "This is the type of {@link AppModel.id} - used to ensure its correct when used in routers etc."
			},
...
			"DataResponseBody_Array_ReadResponseContent_BookingModelDTO.never.never__.Booking.FoundResponseMessageT_": {
				"description": "Represents the response body of a data response",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ReadResponseContent_BookingModelDTO.never.never_"
						},
						"type": "array"
					},
					"message": {
						"$ref": "#/components/schemas/FoundResponseMessageT"
					},
					"metadata": {
						"$ref": "#/components/schemas/MetaData_Booking_"
					}
				},
				"required": [
					"data",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookingReadManyResponse": {
				"$ref": "#/components/schemas/DataResponseBody_Array_ReadResponseContent_BookingModelDTO.never.never__.Booking.FoundResponseMessageT_",
				"description": "Possible responses for {@link BookingController.getAllBookings}"
			},
			"BookingNotFoundResponseBody": {
				"$ref": "#/components/schemas/NotFoundResponseBody",
				"description": "Response type for this controllers object not found response"
			},
...
			"BookingReadResponse": {
				"$ref": "#/components/schemas/DataResponseBody_ReadResponseContent_BookingModelDTO.never.never_.Booking.FoundResponseMessageT_",
				"description": "Possible responses for {@link BookingController.getBooking}"
			},
...
			"BookingCreateResponse": {
				"$ref": "#/components/schemas/DataResponseBody_CreateResponseContent_BookingModelDTO.never.never_.Booking.CreatedResponseMessageT_",
				"description": "Possible responses for {@link BookingController.createBooking}"
			},
...
			"BookingCreateRequest": {
				"$ref": "#/components/schemas/CreateRequestBody_BookingModelDTO.bookingType.never_",
				"description": "Request body format for {@link BookingController.createBooking}'s `booking` body parameter"
			},
...
			"BookingUpdateResponse": {
				"$ref": "#/components/schemas/DataResponseBody_UpdateResponseContent_BookingModelDTO.never.never_.Booking.UpdatedResponseMessageT_",
				"description": "Possible responses for {@link BookingController.updateBooking}"
			},
...
			"BookingUpdateRequest": {
				"$ref": "#/components/schemas/UpdateRequestBody_BookingModelDTO.bookingType.never_",
				"description": "Request body format for {@link BookingController.updateBooking}'s `booking` body parameter"
			},
...
			"BookingDeleteResponse": {
				"$ref": "#/components/schemas/DeleteResponseBody_Booking_",
				"description": "Possible responses for {@link BookingController.deleteBooking}"
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "api_key",
				"in": "query"
			},
			"oauth2": {
				"type": "oauth2",
				"flows": {
					"password": {
						"tokenUrl": "/auth/login",
						"scopes": {
							"admin": "modify things",
							"staff": "modify some things",
							"user": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": ""
	},
	"paths": {
...
		"/bookings": {
			"get": {
				"operationId": "GetAllBookings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingReadManyResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of all bookings, matching the query parameters if provided.",
				"tags": [
					"Booking"
				],
				"security": [
					{
						"api_key": []
					},
					{
						"oauth2": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"description": "- Retrieve records starting from this offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- Retrieve up to this many records",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- Order records by this attribute",
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderByFieldValues"
						}
					},
					{
						"description": "- Stringified JSON where the keys represent a property of a booking and values represent a text based filter for the property",
						"in": "query",
						"name": "filters",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingCreateResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Creates a new booking",
				"tags": [
					"Booking"
				],
				"security": [
					{
						"api_key": []
					},
					{
						"oauth2": [
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- The posted body object, containing the creation attributes of BookingModel",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookingCreateRequest",
								"description": "- The posted body object, containing the creation attributes of BookingModel"
							}
						}
					}
				}
			}
		},
		"/bookings/{id}": {
			"get": {
				"operationId": "GetBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingReadResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing booking with a matching id",
				"tags": [
					"Booking"
				],
				"security": [
					{
						"api_key": []
					},
					{
						"oauth2": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"description": "- The identifier of the booking to get",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/AppModelIdT"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingUpdateResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Update a booking with a matching id",
				"tags": [
					"Booking"
				],
				"security": [
					{
						"api_key": []
					},
					{
						"oauth2": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"description": "- The identifier of the booking to update",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/AppModelIdT"
						}
					}
				],
				"requestBody": {
					"description": "- The posted body object, containing some or all modifiable attributes of BookingModel",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookingUpdateRequest",
								"description": "- The posted body object, containing some or all modifiable attributes of BookingModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingDeleteResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Delete the booking with the matching id",
				"tags": [
					"Booking"
				],
				"security": [
					{
						"api_key": []
					},
					{
						"oauth2": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"description": "- The identifier of the booking to delete",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/AppModelIdT"
						}
					},
					{
						"description": "- If true the booking will actually be deleted from the database rather than a soft delete (handled by sequelize)",
						"in": "query",
						"name": "force",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"basePath": "/",
...
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "api_key",
			"in": "query"
		},
		"oauth2": {
			"type": "oauth2",
			"tokenUrl": "/auth/login",
			"flow": "password",
			"scopes": {
				"admin": "modify things",
				"staff": "modify some things",
				"user": "read things"
			}
		}
	}
}